[kafka/3.x] DKAFC-5098: Added a metric in controller to find URPs caused by a broker.
Summary:
Added a metric in controller to find URPs caused by a broker. This helps get the brokers which are lagging the replication for the topic partitions.

Metric is: kafka.controller:type=UnderReplicatedPartitionMetrics,name=UrpsCausedByBroker tags:<broker_id=>

Added UrpsCausedByBroker metric for all the live brokers even when they do not have any.
This makes sure that there is always a metric for a live broker and its value will be zero if there are no URPs.

Test Plan: Added UTs

Reviewers: kchandraprakash, abhijeek

Reviewed By: kchandraprakash

JIRA Issues: DKAFC-5098

Differential Revision: https://code.uberinternal.com/D15260357