Removed FileChannel#transferFrom call and using Files.copy while fetching indexes from RSM to build index cache.
FileChannel#transferFrom call may be efficient as it goes through a fast path of transferring directly from the source channel into the filesystem cache. But if it goes through non-fast path then it expects the inputStream to always have available bytes. This is an unnecessary restriction on RemoteStorageManager to always return InputStream to have available bytes till the end.