[kafka/3.x] Add security providers
Summary:
Add security providers.

Kafka-security v0.0.4 contains the necessary changes required in authorizer. 3.x build requires us to implement KafkaPrincipalSerde interface. Until the testing is complete we will be using this jar.

We will also be adding caffeine lib dependency explicitly in core module, as kafkasecurity jar imports a version of caffeine that is different and has breaking api changes used in few tests. Hence we are adding a dependency to avoid any transitive dependency issue. We can remove it after the monorepo upgrades to supporting version.

Test Plan:
Authorizer class seamlessly added and is working as expected too. There is a reimplementation of authorizer which is pushed to 0.0.4 security jar which can be used for further testing , we will update it to use the latest jar once testing is complete.
{F970028823}

Reviewers: #ldap_kafka_admins, kchandraprakash

Reviewed By: #ldap_kafka_admins, kchandraprakash

Subscribers: kchandraprakash

JIRA Issues: DKAFC-5180

Differential Revision: https://code.uberinternal.com/D15089281