KAFKA-15859 Make RemoteListOffsets call an async operation (#16602)
Summary:
This is the part-2 of the KIP-1075

To find the offset for a given timestamp, ListOffsets API is used by the client. When the topic is enabled with remote storage, then we have to fetch the remote indexes such as offset-index and time-index to serve the query. Also, the ListOffsets request can contain the query for multiple topics/partitions.

The time taken to read the indexes from remote storage is non-deterministic and the query is handled by the request-handler threads. If there are multiple LIST_OFFSETS queries and most of the request-handler threads are busy in reading the data from remote storage, then the other high-priority requests such as FETCH and PRODUCE might starve and be queued. This can lead to higher latency in producing/consuming messages.

In this patch, we have introduced a delayed operation for remote list-offsets call. If the timestamp need to be searched in the remote-storage, then the request-handler threads will pass-on the request to the remote-log-reader threads. And, the request gets handled in asynchronous fashion.

Covered the patch with unit and integration tests.

Apach-Reviewers: Satish Duggana <satishd@apache.org>, Luke Chen <showuon@gmail.com>, Chia-Ping Tsai <chia7712@gmail.com>
Revision: [[ https://github.com/apache/kafka/commit/344d8a60af8ba7e4ae646501494b17b6987a95d3 | 344d8a60af8ba7e4ae646501494b17b6987a95d3 ]]

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd

Subscribers: satishd

JIRA Issues: DKAFC-5375

Differential Revision: https://code.uberinternal.com/D15391957