[kafka/3.x] Make Remote log metadata events backward compatible with 2.9 build
Summary:
In upstream, `customMetadata` field was introduced when the tiered storage was not GAed, so it went with backward incompatible change.

In 2.9-tiering-dev branch, we marked that field as tagged to ensure compatibilty since we were using the tiered storage feature in 2.9 branch. This change ensures that the events in 2.9 and 3.9 are compatible with each other.

Reference: D14105647

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd

Subscribers: satishd

Revert Plan: git revert

JIRA Issues: DKAFC-5179

Differential Revision: https://code.uberinternal.com/D15599639