KAFKA-18199; Fix size calculation for nullable tagged structs (#18127)
Summary:
When a struct field is tagged and nullable, it is serialized as
{ varint tag; varint dataLength; nullable data }, where
nullable is serialized as
{ varint isNotNull; if (isNotNull) struct s; }. The length field
includes the is-not-null varint.

This patch fixes a bug in serialization where the written value of
the length field and the value used to compute the size of the length
field differs by 1. In practice this has no impact unless the
serialized length of the struct is 127 bytes, since the varint encodings
of 127 and 128 have different lengths (0x7f vs 0x80 01).

Apache-Reviewers: David Jacot <djacot@confluent.io>

Cherrypick: https://github.com/apache/kafka/commit/dcb16bc6977a0a6d642275ba9f8248ffcab5576b

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash, ureview

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-6097

Differential Revision: https://code.uberinternal.com/D17861937