DKAFC-5681: Fix hdfs remote-log segment dump tool
Summary:
While uploading the data to remote storage, we combine the segment files together with the index and snapshot files to avoid storing small files in Hadoop. The below files get merged into a single file:

1. Segment (.log)
2. Offset index (.index)
3. Timestamp index (.timeindex)
4. Producer Snapshot (.snapshot) and
5. Transaction Index (.txnindex)

The DumpHDFSRemoteLogSegment script is used to download that single file from HDFS and split it into the segment and index files. Once the files are split, then we can use the regular `kafka-dump-log.sh` script to decode the files further.

NOTE: The reason for not adding this tool inside the source folder is to avoid some `core` module dependencies in the `remote-storage-managers` module. So, updated the comment in the file for quick reference to run the script.

**Why do we need this diff? **
In the RemoteLogSegmentMetadata, a new validation is introduced to ensure that endOffset > baseOffset, so the script fails in 3.9 branch.

Test Plan: Tested the patch in h20aperf-phx cluster

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, shonali, abhijeek

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, shonali, abhijeek

Subscribers: abhijeek

Tags: #has_java

JIRA Issues: DKAFC-5109, DKAFC-5681

Differential Revision: https://code.uberinternal.com/D16409437