Use Netty PooledByteBuf for reducing GC activity
Summary:
Forward port for D16468973 to 3.9.x-dev branch

HDFS Cache is used to cache the data read from HDFS. CachedInputStream takes care of caching this data and returning from cache if it exists, instead of fetching it from HDFS.

When reading from the HDFS input stream, CachedInputStream currently creates a new byte array of the desired length and passes it to the HDFS InputStream.
This data is cached and also used to copy into the caller-supplied byte array during the read call.

This byte array has a very short life. It is not returned to the caller (only used for copying to the supplied byte array). It is also evicted from the LRU cache when newer
entries are put into the cache.

Creating a large number of short-lived objects causes GC on the server. In this diff, we are introducing Netty's PooledByteBuf to pool the byte array to avoid creating new objects which will soon be GC'd.

Test Plan: Unit Tests

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash

Subscribers: kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-5714

Differential Revision: https://code.uberinternal.com/D16596091