DKAFC-5685: Fix the remote-log-metadata events incompatibility b/w 2.9 and 3.9 branches. (3.9)
Summary:
This change is done so that remote-log-metadata events written using the 2.9 binary can be decodable using the 3.9 binary. The record serde uses AbstractApiMessageSerde to serialize and de-serialize the records.

Due to an incompatible change made in https://github.com/apache/kafka/pull/11010, this change is required to support remote-log-metadata events from 3.9 build;

NOTE: THIS IS A TEMPORARY FIX and will be removed in the next release once all the events written to __remote_log_metadata topic are updated with the new frame version. Since the FRAME version is hard-coded in the code, not writing the unit test for this change.

Test Plan: Verified the changes by running the 2.9 and 3.9 servers locally and ensured that the remote-log-metadata topic events are compatible.

Reviewers: satishd, shonali, abhijeek, #ldap_kafka_admins

Reviewed By: satishd, #ldap_kafka_admins

Tags: #has_java

Revert Plan: git revert

JIRA Issues: DKAFC-5685

API Changes: NA

Monitoring and Alerts: NA

Differential Revision: https://code.uberinternal.com/D16448975