Revert "Use Netty PooledByteBuf for reducing GC activity"
Summary:
This reverts commit 0b9cf99723f2757ee5a07344e9b33e816c84f9b9.

Netty's pooled buffer pool doesn't work well when different threads could allocate and release the bytebuf.
For the pool to work efficiently, the same thread should allocate and release it back to the pool

Because of the above behaviour, we continue to see allocations happening in the steady state. Since our use case is simple,
where the allocated memory is required for the cache and the cache entry size is fixed, we can build a custom buffer pool
that allocates memory of fixed size.

We added a custom Buffer pool in D17189563 which will be ported to 3.9.x branch in a follow-up diff

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash, satishd

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash, satishd

Subscribers: satishd, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-5714

Differential Revision: https://code.uberinternal.com/D17314867