Use custom Buffer Pool for LRU Cache
Summary:
Netty's pooled buffer pool doesn't work well when different threads could allocate and release the bytebuf.
For the pool to work efficiently, the same thread should allocate and release it back to the pool

Because of the above behaviour, we continue to see allocations happening in the steady state. Since our use case is simple,
where the allocated memory is required for the cache and the cache entry size is fixed, we can build a custom buffer pool
that allocates memory of fixed size.

For very low cases, where the complete allocated memory will not be used (only when
we are reading the last chunk of the HDFS file), this wastage of memory (< 1%) is acceptable (1 chunk of 125 chunks per file - since
cache entry size is 4 MB, and the  file size is 500 MB)

D17314867 - reverts Netty ByteBuf changes and this diff adds the Custom BufferPool

Test Plan: - Unit Tests

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-5933

Differential Revision: https://code.uberinternal.com/D17315597