DKAFC-5962: Configure linger.ms to have better batching in RLMM metadata producer
Summary:
In Kafka v4.0, the recommended linger.ms value for lossless producer is 5 ms to reduce the CPU cycles on the broker. As the broker work amplifies with smaller batches.

See:

1. https://www.linkedin.com/pulse/benefits-lingerms5-apache-kafka-jason-taylor-pk3ce/
2. https://cwiki.apache.org/confluence/display/KAFKA/KIP-1030%3A+Change+constraints+and+default+values+for+various+configurations#KIP1030:Changeconstraintsanddefaultvaluesforvariousconfigurations-#5Config(s):linger.ms

This will enable better batching as each record value size ~100 bytes, so in concurrent setup, the producer should be able to batch ~100 events.

Test Plan: Deployed the patch in the kafka-h20aperf-phx cluster

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd

Subscribers: satishd

Tags: #has_java

JIRA Issues: DKAFC-5962

Differential Revision: https://code.uberinternal.com/D17442601