DKAFC-6114: DescribeLogDirs API should include partition size in the remote storage.
Summary:
DescribeLogDirs API should also return the value of:

`RemoteLogSize`: Size of the log segments in the remote storage. This includes only the valid log segments size that are in COPY_SEGMENT_FINISHED state. Dangling log segments won't be counted.
`OnlyLocalLogSize`: Kafka uploads the segments aggressively to remote storage. There can be few segments that are common in both remote and local storage. This attribute will only provide the size of log segments that are yet to be uploaded to remote.

`PartitionSize`: The existing field/attribute is untouched as it might be used by monitoring tools such as Cruise control to balance the disk of brokers.

The additional two fields: `RemoteLogSize` and `OnlyLocalLogSize` help us to compute the exact cost of the topic when remote storage is enabled on the topic.

cherrypick: https://code.uberinternal.com/D17925963

Test Plan:
UT
Verified the patch by running LocalTieredStorage and ensured the output for all the cases.

Sample output:
case: All the passive segments exists in both local and remote log except the active segment:
```
{
    "broker": 1,
    "logDirs": [
        {
            "partitions": [
                {
                    "onlyLocalLogSize": 91158160,
                    "partition": "mango-0",
                    "size": 1244480000,
                    "remoteLogSize": 1153321840,
                    "offsetLag": 0,
                    "isFuture": false
                }
            ],
            "error": null,
            "logDir": "/tmp/kafka-logs1"
        }
    ]
}
```

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, amang

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, amang

Tags: #has_java

JIRA Issues: DKAFC-6114

Differential Revision: https://code.uberinternal.com/D17983883