DKAFC-5974: Gracefully handle error while building remoteLogAuxState in ReplicaFetcherThread
Summary:
Error handling has been improved while building the remote-log-auxiliary state when a follower node with an empty disk begin to synchronise with the leader. If the topic has remote storage enabled, then the ReplicaFetcherThread attempt to build the remote-log-auxiliary state. Note that the remote-log-auxiliary state gets invoked only when the leader-log-start-offset is non-zero and leader-log-start-offset is not equal to leader-local-log-start-offset.

When the LeaderAndISR request is received, then the ReplicaManager#becomeLeaderOrFollower invokes 'makeFollowers' initially, followed by the RemoteLogManager#onLeadershipChange call. As a result, when ReplicaFetcherThread initiates the RemoteLogManager#fetchRemoteLogSegmentMetadata, the partition may not have been initialized at that time.

Introducing a new RetriableRemoteStorageException requires a KIP as it is a public API change, so wrapped the IllegalStateException in RemoteStorageException to gracefully handle the error.

Test Plan:
- Unit test
- Locally reproduced the issue to verify the patch:
```
[2025-07-19 13:23:33,684] INFO [ReplicaFetcher replicaId=1, leaderId=0, fetcherId=0] Could not build remote log auxiliary state for orange-1 due to error: RemoteLogManager is not ready for partition: orange-1 (kafka.server.ReplicaFetcherThread)
...
[2025-07-19 13:23:33,685] INFO [ReplicaFetcher replicaId=1, leaderId=0, fetcherId=0] Could not build remote log auxiliary state for orange-0 due to error: RemoteLogManager is not ready for partition: orange-0 (kafka.server.ReplicaFetcherThread)
```

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, satishd

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd

Subscribers: satishd

Tags: #has_java

JIRA Issues: DKAFC-5974

Differential Revision: https://code.uberinternal.com/D18416803