Add basic implementation for PrefetchEnabledHDFSRemoteStorageManager
Summary:
In this diff, we are adding a new implementation of RSM called PrefetchEnabledHDFSRemoteStorageManager
that wraps the HDFSRemoteStorageManager and delegates most of the operations to it, except for fetch segment
calls, where it checks if the required data is already prefetched and therefore can be used. If the data is not prefetched,
it will delegate the call to HDFSRemoteStorageManager.

Test Plan: Unit Tests

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, satishd, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-6223

Differential Revision: https://code.uberinternal.com/D18265741