RemoteDataPrefetcher Implementation
Summary:
In this diff, we add implementation for RemoteDataFetcher that looks up
the next segment to prefetch based on the current segment being read, and
delegates the task to download the required segment to PrefetchSegmentManager.

It also retrieves the segment from the PrefetchSegmentManager to serve any segments
that are prefetched.

Test Plan: Unit Tests

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, satishd, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-6223

Differential Revision: https://code.uberinternal.com/D18266133