Add PrefetchSegmentManager implementation
Summary:
In this diff, we add the implementation for PrefetchSegmentManager, which is responsible
for downloading the segments for remote storage when required. It saves the downloaded
segment on the local disk and keeps it reference in a cache. It serves the segments when
required.

PrefetchSegmentManager is also responsible for expiring the downloaded segments to prevent
the local disk from growing in size.

Test Plan: Unit Tests

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, satishd, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-6229

Differential Revision: https://code.uberinternal.com/D18266331