Add FileSystemManager for managing filesystems used by HDFS RemoteStorageManager
Summary:
In this diff, we are extracting the HDFS/OCS FileSystem management logic to a separate class.
This change is required because DataFetcher used for prefetching will also need to interact with these
filesystems. Extracting this logic into FileSystemManager and use it for DataFetcher and HDFSRemoteStorageManager
helps avoid duplicate logic and also have a single instance of a FileSystem per hadoop configuration

Test Plan: Existing Unit Tests

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, satishd, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-6227

Differential Revision: https://code.uberinternal.com/D18355427