HDFSDataFetcher implementation
Summary:
In this diff, we add the implementation for HDFSDataFetcher. It uses hadoop configuration
that allows it to quickly download a segment from the remote storage (HDFS or OCS)
In the current diff, we only add HDFS based configuration (to use hedged reads). In a future diff,
we will add OCS based configuration.

Test Plan: Unit Tests

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, satishd, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-6227

Differential Revision: https://code.uberinternal.com/D18360501