KAFKA-19425: Stop the server when fail to initialize to avoid local segment never got deleted. (#20007)
Summary:
We found that one broker's local segment on disk never gets removed no matter how long it is stored. The disk always keeps growing.

![image](https://github.com/user-attachments/assets/42129bb6-7d07-481b-923f-971da3ab12da)
note: Partition 2's node is the exception node.

After the trouble shooting, it was found if one broker is very slow to startup, it will cause the TopicBasedRemoteLogMetadataManager#initializeResources's fail sometimes
(it meets the expectation due to the server is not ready as fast as expected). Thus it won't stop the server so that the server still runs just with some exception log but not shutdown.  It won't upload to remote as the local segment never gets deleted.

So proposed the change to shut down the broker to avoid the silent critical error that caused the disk growing forever.

Apache Reviewers: Kamal Chandraprakash <kamal.chandraprakash@gmail.com>, Luke
 Chen <showuon@gmail.com>

Test Plan: UT

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash

Subscribers: kchandraprakash

Tags: #has_java

Differential Revision: https://code.uberinternal.com/D18637589