Quota Manager integration for prefetch
Summary:
In this diff, we are introducing quotas for the prefetch feature of Tiered Storage.
This is required because when prefetch is enabled, the broker can download
data from the remote storage very quickly and cause degradation of remote storage/broker.

We use the existing quota manager and plug it in for prefetch. The quotas are static currently
and will be made dynamic in a future diff.

Test Plan:
Unit Tests
Tested on dev cluster

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, satishd, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, satishd, kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-6298

Differential Revision: https://code.uberinternal.com/D18897753