Port fetchlookback metrics to 3.9.x
Summary:
This is a port of the fetch lookback metric implementation done in kafka v2.9.x. The idea is to capture the maximum lookback period of the fetch calls happening for a given topic.
It is direct logical translation of these 3 diffs:

  # https://code.uberinternal.com/D10593727
  # https://code.uberinternal.com/D11103303
  # https://code.uberinternal.com/D18654907

Test Plan: UTs

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, kchandraprakash

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, kchandraprakash

Subscribers: kchandraprakash

Tags: #has_java

JIRA Issues: DKAFC-6255

Differential Revision: https://code.uberinternal.com/D19274939