DKAFC-6527: Don't mark the remote-read requests as failed when ReplicaNotAvailableException thrown
Summary:
When the remote log metadata is not ready for a partition during broker restarts, then the ReplicaNotAvailableException gets thrown and any remote-read requests to those partitions are marked as failed.

This creates noise for the oncall. The underlying remote storage is healthy, but the metadata is not yet ready. The consumer/client retries those requests. So, not marking those requests as failed when hitting ReplicaNotAvailableException.

Test Plan: UT

Reviewers: #ldap_kafka_admins, #ldap_kafka_core_reviewers, ureview, amang

Reviewed By: #ldap_kafka_admins, #ldap_kafka_core_reviewers, amang

Tags: #has_java

JIRA Issues: DKAFC-6527

Differential Revision: https://code.uberinternal.com/D19521849