DKAFC-6647: Use circuit breaker to handle remote copy/delete errors (#17)
Summary:
- Use resilience4j circuit breaker to handle remote copy/delete errors.
Updated HDFSRemoteStorageManager to skip operations when the circuit is
open for copy and delete operations. Note that the copy / delete
operations are scheduled to run once in 30 seconds for a partition.
- Use sleep/wait logic to handle the remote fetch errors. We cannot
return the call immediately as the consumer might send another request
to the broker and utilize the request handler threads.
- Added unit tests
- T3-DKAFC-6647

doc: https://resilience4j.readme.io/docs/circuitbreaker