DKAFC-6695: Implement ISR BlockList as dynamic broker config (#25)
## Summary
In this PR, we are implementing ISR block list as a dynamic broker
config.
Previously, the feature was implemented using zookeeper, where the admin
has to make a zookeeper entry to add nodes to ISR block list. With
zookeeper getting deprecated (because of KRaft), we are reimplementing
the feature to remove
zookeeper dependency. The feature going forward will work with KRaft
mode. For backward compatibility, we will continue to support zookeeper
based feature, until all zookeeper based usages are migrated to using
dynamic broker config.

Admins are expected to set the dynamic config as a cluster default
instead of a specific broker default.


## Test Plan
- Unit Tests
- Tested on a dev cluster

```
[udocker@/home/udocker/odin-kafka #]bin/kafka-topics.sh --bootstrap-server localhost:9092 --describe --under-replicated-partitions
[udocker@/home/udocker/odin-kafka #]bin/kafka-configs.sh --bootstrap-server localhost:9092 --entity-type brokers --entity-default --alter --add-config isr.block.list=10036
Completed updating default config for brokers in the cluster.
[udocker@/home/udocker/odin-kafka #]bin/kafka-topics.sh --bootstrap-server localhost:9092 --describe --under-replicated-partitions
	Topic: demo	Partition: 28	Leader: 10032	Replicas: 10036,10032,10027	Isr: 10027,10032	Elr: N/A	LastKnownElr: N/A
```



## Jira Issues
T3-DKAFC-6695